api_server/requirements.txt
fastapi[standard]
pandas
passlib[bcrypt]
psycopg2
pyjwt
sqlalchemy
sqlalchemy-utils
typing_extensions